Feat company logo

+++ back/app/controllers/company_controller.ts
@@ -7,7 +7,7 @@ import {
     getCompanyValidator,
     searchCompaniesValidator,
     updateCompanyValidator,
-    deleteCompaniesValidator,
+    deleteCompanyValidator,
     confirmCompanyValidator,
 } from '#validators/company';
 import axios from 'axios';
@@ -21,11 +21,17 @@ import NominatimService from '#services/nominatim_service';
 import { parsePhoneNumberFromString, PhoneNumber } from 'libphonenumber-js';
 import cache from '@adonisjs/cache/services/main';
 import PaginatedCompanies from '#types/paginated/paginated_companies';
-import SerializedCompany from '#types/serialized/serialized_company';
 import CountryService from '#services/country_service';
 import FileService from '#services/file_service';
 import OpenAiApiService from '#services/open_ai_service';
 import { OpenAiCompanyVerificationResult } from '#types/open-ai/open_ai_company_verification_result';
+import app from '@adonisjs/core/services/app';
+import File from '#models/file';
+import path from 'node:path';
+import FileTypeEnum from '#types/enum/file_type_enum';
+import SlugifyService from '#services/slugify_service';
+import SerializedCompanySuperLight from '#types/serialized/serialized_company_super_light';
+import StringService from '#services/string_service';
 
 @inject()
 export default class CompanyController {
@@ -34,10 +40,12 @@ export default class CompanyController {
         private readonly nominatimService: NominatimService,
         private readonly countryService: CountryService,
         private readonly fileService: FileService,
-        private readonly openAiApiService: OpenAiApiService
+        private readonly openAiApiService: OpenAiApiService,
+        private readonly slugifyService: SlugifyService,
+        private readonly stringService: StringService
     ) {}
 
-    public async getFromSiret({ request, response, i18n }: HttpContext): Promise<void> {
+    public async getFromSiret({ request, response, i18n }: HttpContext) {
         const { siret } = await getCompanyFromSiretValidator.validate(request.params());
 
         try {
@@ -65,8 +73,8 @@ export default class CompanyController {
         }
     }
 
-    public async create({ request, response, user, language, i18n }: HttpContext): Promise<void> {
-        const { siret, name, address: inputAddress, postalCode, city, complement, countryCode, email, phoneNumber: inputPhoneNumber } = await request.validateUsing(createCompanyValidator);
+    public async create({ request, response, user, i18n }: HttpContext) {
+        const { siret, name, address: inputAddress, postalCode, city, complement, countryCode, email, phoneNumber: inputPhoneNumber, logo } = await request.validateUsing(createCompanyValidator);
 
         let company: Company | null = await this.companyRepository.findOneBy({ siret });
         if (company) {
@@ -110,7 +118,6 @@ export default class CompanyController {
             latitude: data.latitude,
             longitude: data.longitude,
         });
-        await address.refresh();
 
         company = await Company.create({
             siret,
@@ -119,7 +126,6 @@ export default class CompanyController {
             phoneNumber: phoneNumber?.format('E.164'),
             addressId: address.id,
         });
-        await company.refresh();
 
         await CompanyAdministrator.create({
             role: CompanyAdministratorRoleEnum.CEO,
@@ -127,25 +133,30 @@ export default class CompanyController {
             userId: user.id,
         });
 
-        await Promise.all([company.load('address'), company.load('equipments'), cache.deleteByTag({ tags: ['companies'] })]);
+        if (logo) {
+            company = await this.updateLogo(company, logo);
+        }
+
+        await Promise.all([cache.deleteByTag({ tags: [`companies:administrator:${user.id}`] }), company.save()]);
+        await Promise.all([company.load('address'), company.load('logo')]);
 
         return response.created({
             message: i18n.t('messages.company.create.success', { companyName: company.name }),
-            company: company.apiSerializeLight(language),
+            company: company.apiSerializeSuperLight(),
         });
     }
 
-    public async getAll({ request, response, user, language }: HttpContext): Promise<void> {
+    public async getAll({ request, response, user, language }: HttpContext) {
         const { query, page, limit, sortBy: inputSortBy } = await request.validateUsing(searchCompaniesValidator);
 
         return response.ok(
             await cache.getOrSet({
                 key: `companies:${user.id}:query:${query.toLowerCase()}:page:${page}:limit:${limit}:sortBy:${inputSortBy}`,
-                tags: ['companies'],
+                tags: [`companies:administrator:${user.id}`],
                 ttl: '1h',
                 factory: async (): Promise<PaginatedCompanies> => {
                     const [field, order] = inputSortBy.split(':');
-                    const sortBy = { field: field as keyof Company['$attributes'], order: order as 'asc' | 'desc' };
+                    const sortBy = { field: this.stringService.toSnakeCase(field) as keyof Company['$attributes'], order: order as 'asc' | 'desc' };
 
                     return await this.companyRepository.getProfileCompanies(user, language, query.toLowerCase(), page, limit, sortBy);
                 },
@@ -153,27 +164,22 @@ export default class CompanyController {
         );
     }
 
-    public async delete({ request, response, i18n, user }: HttpContext): Promise<void> {
-        const { companies } = await request.validateUsing(deleteCompaniesValidator);
+    public async delete({ request, response, i18n, user }: HttpContext) {
+        const { companyId } = await request.validateUsing(deleteCompanyValidator);
 
-        const statuses: { isDeleted: boolean; name?: string; id: string }[] = await this.companyRepository.delete(companies, user);
+        const statuses: { isDeleted: boolean; name?: string; id: string }[] = await this.companyRepository.delete([companyId], user);
+        const status: { isDeleted: boolean; name?: string; id: string } = statuses[0];
 
-        return response.ok({
-            messages: await Promise.all(
-                statuses.map(async (status: { isDeleted: boolean; name?: string; id: string }): Promise<{ id: string; message: string; isSuccess: boolean }> => {
-                    if (status.isDeleted) {
-                        await cache.deleteByTag({ tags: ['companies', `company:${status.id}`] });
-                        return { id: status.id, message: i18n.t(`messages.company.delete.success`, { name: status.name }), isSuccess: true };
-                    } else {
-                        return { id: status.id, message: i18n.t(`messages.company.delete.error.default`, { id: status.id }), isSuccess: false };
-                    }
-                })
-            ),
-        });
+        if (status.isDeleted) {
+            await cache.deleteByTag({ tags: [`companies:administrator:${user.id}`, `company:${status.id}`] });
+            return response.ok({ messages: [{ message: i18n.t(`messages.company.delete.success`, { name: status.name }), isSuccess: true }] });
+        } else {
+            return response.ok({ messages: [{ message: i18n.t(`messages.company.delete.error.default`, { id: status.id }), isSuccess: false }] });
+        }
     }
 
-    public async update({ request, response, i18n, language, user }: HttpContext) {
-        const { siret, name, address: inputAddress, postalCode, city, complement, countryCode, email, phoneNumber: inputPhoneNumber } = await request.validateUsing(updateCompanyValidator);
+    public async update({ request, response, i18n, user }: HttpContext) {
+        const { companyId, name, address: inputAddress, postalCode, city, complement, countryCode, email, phoneNumber: inputPhoneNumber, logo } = await request.validateUsing(updateCompanyValidator);
 
         const country: Country | undefined = CountryList.default.findOneByCountryCode(countryCode);
         if (!country) {
@@ -182,7 +188,7 @@ export default class CompanyController {
             });
         }
 
-        const company: Company = await this.companyRepository.getFromUser(siret, user);
+        let company: Company = await this.companyRepository.getFromUser(companyId, user);
 
         const fullAddress: string = `${inputAddress}, ${postalCode} ${city}, ${country.name}`;
         const data: { latitude: number; longitude: number } | null = await this.nominatimService.getFromAddress(fullAddress);
@@ -217,14 +223,28 @@ export default class CompanyController {
             company.phoneNumber = phoneNumber.format('E.164');
         }
 
-        await Promise.all([company.save(), company.address.save(), cache.deleteByTag({ tags: ['companies', `company:${company.id}`] })]);
+        if (logo) {
+            company = await this.updateLogo(company, logo);
+        }
 
-        return response.ok({ company: company.apiSerializeLight(language), message: i18n.t('messages.company.update.success', { name }) });
+        await Promise.all([
+            company.save(),
+            company.address.save(),
+            cache.deleteByTag({ tags: [`companies:administrator:${user.id}`] }),
+            cache.set({
+                key: `company:${company.id}`,
+                tags: [`company:${company.id}`],
+                ttl: '1h',
+                value: company.apiSerialize(),
+            }),
+        ]);
+
+        return response.ok({ company: company.apiSerializeLight(), message: i18n.t('messages.company.update.success', { name }) });
     }
 
-    public async get({ request, response, i18n, language }: HttpContext): Promise<void> {
+    public async getOne({ request, response, i18n, user }: HttpContext) {
         const { companyId } = await getCompanyValidator.validate(request.params());
-        const company: Company | null = await this.companyRepository.findOneBy({ id: companyId }, ['administrators']);
+        const company: Company | null = await this.companyRepository.getFromUser(companyId, user);
         if (!company) {
             return response.notFound({ error: i18n.t('messages.company.get.error.not-found') });
         }
@@ -234,8 +254,8 @@ export default class CompanyController {
                 key: `company:${company.id}`,
                 tags: [`company:${company.id}`],
                 ttl: '1h',
-                factory: (): SerializedCompany => {
-                    return company.apiSerialize(language);
+                factory: (): SerializedCompanySuperLight => {
+                    return company.apiSerializeSuperLight();
                 },
             }),
             countries: await cache.getOrSet({
@@ -249,7 +269,7 @@ export default class CompanyController {
         });
     }
 
-    public async confirm({ request, response, user, i18n, language }: HttpContext): Promise<void> {
+    public async confirm({ request, response, user, i18n }: HttpContext) {
         const { companyId, document } = await request.validateUsing(confirmCompanyValidator);
         if (!document.tmpPath) {
             return response.badRequest({ error: i18n.t('messages.company.confirm.error.no-document') });
@@ -282,11 +302,54 @@ export default class CompanyController {
 
         company.enabled = true;
 
-        await Promise.all([company.save(), cache.deleteByTag({ tags: ['companies', `company:${company.id}`] })]);
+        await Promise.all([
+            company.save(),
+            cache.deleteByTag({ tags: [`companies:administrator:${user.id}`] }),
+            cache.set({
+                key: `company:${company.id}`,
+                tags: [`company:${company.id}`],
+                ttl: '1h',
+                value: company.apiSerialize(),
+            }),
+        ]);
 
         return response.ok({
             message: i18n.t('messages.company.confirm.success', { name: company.name }),
-            company: company.apiSerialize(language),
+            company: company.apiSerializeSuperLight(),
         });
+    }
+
+    private async updateLogo(company: Company, logo: any): Promise<Company> {
+        if (company.logoId) {
+            // Physically delete the file
+            this.fileService.delete(company.logo);
+
+            const oldLogo: File = company.logo;
+            company.logoId = null;
+            await company.save();
+            await oldLogo.delete();
+        }
+
+        const originalName: string = logo.clientName;
+        const slugifiedName: string = this.slugifyService.slugify(originalName);
+        const extension: string = path.extname(originalName);
+        const uniqueFilename: string = `${slugifiedName.replace(extension, '')}-${Date.now()}${extension}`;
+
+        const logoPath: string = 'static/company-logo';
+        const fullPath: string = app.makePath(logoPath);
+
+        await logo.move(fullPath, { name: uniqueFilename });
+        const newLogo: File = await File.create({
+            name: uniqueFilename,
+            path: `${logoPath}/${uniqueFilename}`,
+            extension,
+            mimeType: `${logo.type}/${logo.subtype}`,
+            size: logo.size,
+            type: FileTypeEnum.COMPANY_LOGO,
+        });
+
+        company.logoId = newLogo.id;
+
+        return company;
     }
 }
